+++ app/config/settings.py
@@ -1,4 +1,5 @@
 """This module contains the settings for the application."""
+
 import os
 from dotenv import load_dotenv
 
@@ -13,3 +14,4 @@ JWT_REFRESH_SECRET_KEY = os.getenv("JWT_REFRESH_SECRET_KEY")
 OM2M_URL = os.getenv("OM2M_URL")
 OM2M_USERNAME = os.getenv("OM2M_USERNAME")
 OM2M_PASSWORD = os.getenv("OM2M_PASSWORD")
+MOBIUS_XM2MRI = os.getenv("MOBIUS_XM2MRI")
