+++ app/utils/om2m_lib.py
@@ -1,158 +1,63 @@
-"""
-This module provides a class Om2m to interact with OM2M API.
-"""
-
-
 import requests
 
 
 class Om2m:
-    """
-    This class provides methods to interact with OM2M API.
-    """
-
-    def __init__(self, username, password, url):
-        """
-        Initializes the class with username, password and url.
-        """
-        self.username = username
-        self.password = password
+    def __init__(self, XM2MRI, url):
+        self.XM2MRI = XM2MRI
         self.url = url
 
-    def create_ae(self, name, path, labels=None, rr=False, timeout=None):
-        """
-        Creates an Application Entity (AE) resource in OM2M.
-        """
-        if labels is None:
-            labels = []
-        data = {"m2m:ae": {"rn": name, "lbl": labels, "rr": rr, "api": name}}
+    def create_ae(self, name, labels=[], rr=False):
+        XM2MORIGIN = "SOrigin" + name
+        data = {
+            "m2m:ae": {
+                "rn": name,
+                "api": "0.2.481.2.0001.001.000111",
+                "lbl": labels,
+                "rr": rr,
+            }
+        }
         headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
+            "X-M2M-RI": self.XM2MRI,
+            "X-M2M-Origin": XM2MORIGIN,
             "Content-Type": "application/json;ty=2",
         }
 
-        print(f"{self.url}/{path}", headers, data)
-        r = requests.post(
-            f"{self.url}/{path}", headers=headers, json=data, timeout=timeout
-        )
-        print(r.status_code, r.text)
+        r = requests.post(self.url, headers=headers, json=data)
         return r.status_code, r.text
 
-    def create_container(self, name, path, labels=None, mni=120, timeout=None):
-        """
-        Creates a node resource in OM2M.
-        """
-        if labels is None:
-            labels = []
+    def create_container(self, name, parent, labels=[], mni=120):
+        XM2MORIGIN = "SOrigin" + parent
+        # Create Node
         data = {"m2m:cnt": {"rn": name, "lbl": labels, "mni": mni}}
         headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
+            "X-M2M-RI": self.XM2MRI,
+            "X-M2M-Origin": XM2MORIGIN,
             "Content-Type": "application/json;ty=3",
         }
 
+        r = requests.post(self.url + "/" + parent, headers=headers, json=data)
+
+        # Create Data Container inside Node
+        data = {"m2m:cnt": {"rn": "Data", "lbl": labels, "mni": mni}}
+
         r = requests.post(
-            f"{self.url}/{path}", headers=headers, json=data, timeout=timeout
+            self.url + "/" + parent + "/" + name, headers=headers, json=data
         )
+
         return r
 
     def create_cin(self, parent, node, con, lbl=None, timeout=None):
-        """
-        Creates a content instance resource in OM2M.
-        """
-        if lbl is None:
-            lbl = [node]
+        XM2MORIGIN = "SOrigin" + parent
         data = {"m2m:cin": {"con": con, "lbl": lbl}}
+
         headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
+            "X-M2M-RI": self.XM2MRI,
+            "X-M2M-Origin": XM2MORIGIN,
             "Content-Type": "application/json;ty=4",
         }
-        url = f"{self.url}/{parent}/{node}"
-        if parent == "" or parent is None:
-            # Now we have direct ri so we also remove /in-name
-            url = self.url.replace("/in-name", "") + f"/{node}"
-        print(url, headers, data)
+
         r = requests.post(
-            url,
-            headers=headers,
-            json=data,
-            timeout=timeout,
-        )
-        return r
-
-    def delete_resource(self, resource_path, timeout=None):
-        """
-        Deletes a resource in OM2M.
-        """
-        headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
-        }
-        print(f"{self.url}/{resource_path}")
-        r = requests.delete(
-            f"{self.url}/{resource_path}",
-            headers=headers,
-            timeout=timeout,
-        )
-        return r
-
-    def get_containers(self, resource_path="", ri=None, all=False, timeout=None):
-        """
-        Retrieves all containers from the OM2M server.
-        """
-        headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
-            "Content-Type": "application/json;ty=3",
-        }
-        url = f"{self.url}/{resource_path}"
-        if ri:
-            url = f"{self.url.replace('/in-name', '')}/{ri}"
-
-        if all:
-            url = f"{url}?rcn=4"
-
-        r = requests.get(url, headers=headers, timeout=timeout)
-        return r
-
-    def get_all_resource(self, resource_path: str, timeout=None):
-        """
-        Gets all resource in OM2M.
-        """
-        headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
-        }
-        r = requests.get(
-            f"{self.url}/{resource_path}?rcn=4",
-            headers=headers,
-            timeout=timeout,
-        )
-        return r
-
-    def get_la_cin(self, resource_path, timeout=None):
-        """
-        Gets latest content instance in OM2M.
-        """
-        headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
-            "Content-Type": "application/json;ty=4",
-        }
-        print(f"{self.url}/{resource_path}/la")
-        r = requests.get(
-            f"{self.url}/{resource_path}/la",
-            headers=headers,
-            timeout=timeout,
-        )
-        return r
-
-    def create_subscription(self, resource_path, rn, nu, nct=2, timeout=None):
-        """
-        Creates a subscription resource in OM2M.
-        """
-        data = {"m2m:sub": {"rn": rn, "nu": nu, "nct": nct}}
-        headers = {
-            "X-M2M-Origin": f"{self.username}:{self.password}",
-            "Content-Type": "application/json;ty=23",
-        }
-        r = requests.post(
-            f"{self.url}/{resource_path}",
+            self.url + "/" + parent + "/" + node + "/Data?rcn=1",
             headers=headers,
             json=data,
             timeout=timeout,
